New tokeniser specifying which read to tokenise against.
This builds a trie to analyse prefixes, but for now proper
optimisation of this eludes me so it just has custom prefix lengths
for PacBio and IonTorrent (this doesn't change the format or decoder).

This gives big saving to data set 03 (PacBio) and minimal savings for
09. However it also has minimal extra space cost on other data sets
and larger cpu/memory requirements.

Examples before and after for 10,000 names taken from different data
sets:

          tokenise  tokenise2  gzip   bzip2  xz     lpaq
01.names  19797     20467      71924  59055  34412  26770
02.names  45996     46490      75554  62650  63532  47319
03.names  82524     59027      93466  72606  78100  54506
05.names  32363     33400      52906  49121  44316  32750
08_names  46730     49098      94418  73189  79408  50236
09.names  38558     37068      53401  39669  38536  28872
10.names  39717     40066      72630  54588  55452  40435
20.names  17325     18090      54038  45386  29508  18447

(09 still stands out as room for improvement vs lpaq)