Improved the evaluation of token compression.
try_comp.sh now handles a "cat" encoder for when the data simply isn't
worth running through an entropy encoder (it's random looking or very
tiny).

Pack_dir take the directory output from try_comp.sh and serialises it
to a single stream so we have a better understanding of actual size.
Unpack_dir reverses this.

Loop.sh is a trivial shell script wrapping all this up to aid testing.