Added RLE as an entropy encoder in its own right.
Added a 32-bit based encoder instead of making the tokenisers emit 4x8
bit descriptors.  This is slow at present due to the shell scripts and
brute force nature, but it's permits us to get the tokenisers
outputting in a more sensible format and put the entropy encoding job
in the correct place.

Added some results for a small test set, to demonstrate a chunked '10k
reads at a time' ratio.  We still need proper looping over the entire
file.