Replaced N_ZERO with N_DIGITS0.
Instead of matching 0* it now matches [0-9]* meaning DIGITS0 can
encode 00127 instead of needing 00 and 127 tokens (N_ZERO and
N_DIGITS).  To do this DIGITS0 has two parameters, length and value,
permitting the leading zero.

This both simplifies the synchronisation of tokens and also removes
the entropy lost in not exploiting correlation of number of leading
zeros and length of remaining digits in data sets that use fixed sized
fields (pretty much all with leading 0s).