Bug fix for encoding names with prefix matches.
Eg:

    name100
    name10

Due to the building order of the trie, this was coming out as an exact
match.  The trie is legacy given we now hard-code patterns for pacbio
and iontorrent, so should be replaced with a simpler hash table; both
faster and more accurate avoiding the prefix issue.  However for
simplicity now I fixed the bug via an additional length check.