Improvements to storing order-1 frequency table.
We store the symbols present in an order-0 analysis and then store all
order-1 frequencies for these order-0 symbols, rather than all the
non-zero order-1 frequencies.  This removes the repetition of the
symbol table but may add many zero frequencies. To mitigate this the
zero-frequencies are now run-length encoded.

Additionally any order-1 table that is sufficiently large (1k+) is
itself passed through the order-0 encoder.

The net effect is usually minor, but on some complex data sets it can
be a considerable reduction in size.  Eg low entropy data, was
15100000 -> 767367 bytes, now 15100000 -> 767185 bytes.

Higher entropy data: was 2373984 to -> 877738 bytes, now 2373984 ->
850398 (a 3% saving).